+++ src/hooks/ready/registerGameSettings.js
@@ -434,6 +434,15 @@ export default function () {
     default: false,
   });
 
+  game.settings.register("ddb-importer", "munching-policy-update-images", {
+    name: "ddb-importer.munching-policy-update-images.name",
+    hint: "ddb-importer.munching-policy-update-images.hint",
+    scope: "player",
+    config: false,
+    type: Boolean,
+    default: true,
+  });
+
   // sets the current tutorial step the user has completed
   // game.settings.register("ddb-importer", "tutorial-step", {
   //   scope: "world",