+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Support new attunement options
 * Better error reporting on config screen
 * Improved character background parsing
+* Option to not replace art on monster munch update
 
 # v 0.1.14